% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{as_sass_layer}
\alias{as_sass_layer}
\title{Sass Bundle to Single Sass Layer}
\usage{
as_sass_layer(x)
}
\description{
Converts a \code{\link[=sass_bundle]{sass_bundle()}} to a single Sass layer object.
}
\details{
This is exported for internal use between packages and should not be used.
Instead, please use \code{\link[=sass_layer]{sass_layer()}} or \code{\link[=sass_bundle]{sass_bundle()}} to construct and manage your sass objects
and \code{\link[=sass]{sass()}} and \code{\link[=as_sass]{as_sass()}} to convert your objects.
}
\keyword{internal}
