% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile Sass to CSS}
\usage{
compile(input, options = sass::opts())
}
\arguments{
\item{input}{Path to .scss or .sass file, or string containing
code to compile.}

\item{options}{Compiler options for Sass. Please specify options
using \code{\link{opts}}.}
}
\value{
One element character vector containing the compiled CSS.
}
\description{
Compile Sass to CSS using libSass.
}
\examples{
compile("foo { margin: 122px * .3; }")

}
