% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sass-deprecated}
\alias{sass-deprecated}
\alias{sass_layer_merge}
\title{Deprecated}
\usage{
sass_layer_merge(...)
}
\description{
Deprecated Sass functions
}
\section{Functions}{
\itemize{
\item \code{sass_layer_merge()}: Please use \code{sass_bundle(...)}

}}
\keyword{internal}
