% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass.R
\name{sass}
\alias{sass}
\title{Compile Sass to CSS}
\usage{
sass(
  input = NULL,
  options = sass_options(),
  output = NULL,
  cache_options = sass_cache_options(),
  write_attachments = NA
)
}
\arguments{
\item{input}{Accepts raw Sass, a named list of variables, or a list of raw Sass and/or named variables.
See \code{\link{as_sass}} and \code{\link{sass_import}} / \code{\link{sass_file}} for more details.}

\item{options}{Compiler options for Sass. Please specify options using
\code{\link{sass_options}}.}

\item{output}{Specifies path to output file for compiled CSS.}

\item{cache_options}{Caching options for Sass. Please specify options using
\code{\link{sass_cache_options}}. Caching is turned off by default for
interactive R sessions, and turned on for non-interactive ones.}

\item{write_attachments}{If the input contains \code{\link{sass_layer}}
objects that have file attachments, and \code{output} is not \code{NULL},
then copy the file attachments to the directory of \code{output}. (Defaults
to \code{NA}, which merely emits a warning if file attachments are present,
but does not write them to disk; the side-effect of writing extra files is
subtle and potentially destructive, as files may be overwritten.)}
}
\value{
If \code{output = NULL}, the function returns a string value
  of the compiled CSS. If the output path is specified, the compiled
  CSS is written to that file and \code{invisible()} is returned.
}
\description{
Compile Sass to CSS using LibSass.
}
\examples{
# raw Sass input
sass("foo { margin: 122px * .3; }")

# list of inputs, including named variables
sass(list(
  list(width = "122px"),
  "foo { margin: $width * .3; }"
))

# import a file
tmp_file <- tempfile()
writeLines("foo { margin: $width * .3; }", tmp_file)
sass(list(
  list(width = "122px"),
  sass_file(tmp_file)
))


# File attachment example: Create a checkboard pattern .png, then
# use it from a sass layer

tmp_png <- tempfile(fileext = ".png")
grDevices::png(filename = tmp_png, width = 20, height = 20,
  bg = "transparent", antialias = "none")
par(mar = rep_len(0,4), xaxs = "i", yaxs = "i")
plot.new()
rect(c(0,0.5), c(0,0.5), c(0.5,1), c(0.5,1), col = "#00000044", border=NA)
dev.off()

layer <- sass_layer(
  rules = ".bg-check { background-image: url(images/demo_checkboard_bg.png) }",
  file_attachments = c("images/demo_checkboard_bg.png" = tmp_png)
)

output_path <- tempfile(fileext = ".css")
sass(layer, output = output_path, write_attachments = TRUE)

}
\seealso{
\url{http://sass-lang.com/guide}
}
