% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass_cache.R
\name{sass_context_cache_dir}
\alias{sass_context_cache_dir}
\title{Return the cache directory for the current context.}
\usage{
sass_context_cache_dir()
}
\description{
Return the cache directory for the current context.
}
\keyword{internal}
