% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass_cache.R
\name{sass_default_cache}
\alias{sass_default_cache}
\title{Get default sass cache object}
\usage{
sass_default_cache()
}
\description{
Get the default sass cache object or directory, for the given context. The
context depends on factors described below.

\code{sass_default_cache()} first checks the \code{sass.cache} option. If it is set to
\code{FALSE}, then this function returns \code{NULL}. If it has been set to a string,
it is treated as a directory name, and this function returns a
\code{sass_file_cache()} object using that directory. If the option has been set
to a \code{sass_file_cache()} object, then it will return that object.

In most cases, this function uses the user's cache directory, by calling
\code{rappdirs::user_cache_dir("R-sass")}.

If this function is called from a Shiny application, it will also look for a
subdirectory named \verb{cache/}. If it exists, it will use a directory named
\verb{cache/sass/} to store the cache.

When running a Shiny application in a user R session, it will not create the
\verb{cache/} subdirectory, but it will use it if present. This scopes the cache
to the application.

With Shiny applications hosted on Shiny Server and Connect, it \emph{will} create
a \verb{cache/sass/} subdirectory, so that the cache is scoped to the application
and will not interfere with another application's cache.
}
\seealso{
\code{\link[=sass_cache_get]{sass_cache_get()}}, \code{\link[=sass]{sass()}}
}
