% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_cache.R
\name{sass_file_cache}
\alias{sass_file_cache}
\title{Create a file cache object}
\usage{
sass_file_cache(dir = NULL, max_size = 40 * 1024^2)
}
\arguments{
\item{dir}{The directory in which to store the cached files. If `NULL` (the
default), then it will be set to a subdirectory of the system temp
directory named `R-sass-cache-username`. Because this cache directory is
not in the R process's temp directory, it will persist longer than the R
process, typically until a system reboot. Additionally, it will be shared
across R processes for the same user on the same system.}

\item{max_size}{The maximum size of the cache. If the cache grows past this
size, the least-recently-used objects will be removed until it fits within
this size.}
}
\description{
This creates a file cache which is to be used by sass for caching generated
.css files.
}
\seealso{
sass_get_default_cache
}
