% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass_cache.R
\name{sass_hash}
\alias{sass_hash}
\title{Returns a hash of the object, including sass_file mtimes}
\usage{
sass_hash(x)
}
\arguments{
\item{x}{A list with sass objects.}
}
\description{
This function returns a hash of the object \code{x}, intended for use in caching.
It recuses into lists, and any \code{\link[=sass_file]{sass_file()}} objects will have the file's
mtime attached as an attribute. This is useful for detecting if the file has
been modified.
}
