% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass-dependencies.R
\name{sass_html_dependencies}
\alias{sass_html_dependencies}
\title{Compile Sass to CSS using LibSass, then return an HTML Dependencies}
\usage{
sass_html_dependencies(
  input,
  name,
  version,
  stylesheet,
  cache_key = NULL,
  options = sass_options(),
  write_attachments = TRUE,
  cache = sass_cache_get(),
  script = NULL,
  meta = NULL,
  head = NULL
)
}
\arguments{
\item{input}{Accepts raw Sass, a named list of variables, or a list of raw
Sass and/or named variables. See \code{\link{as_sass}} and
\code{\link{sass_import}} / \code{\link{sass_file}} for more details.}

\item{name}{Library name}

\item{version}{Library version}

\item{stylesheet}{Absolute file path(s) pointing to stylesheet(s) to include with the dependency}

\item{cache_key}{additional information to include in the \code{sass_hash()} call used to
determine the output directory. This should include any information that could possibly
influence the resulting CSS that isn't already captured by \code{input}. For example, if
\code{input} contains something like \code{"@import sass_file.scss"} you may want to include the
\code{\link[=file.mtime]{file.mtime()}} of \code{sass_file.scss} (or, perhaps, a \code{\link[=packageVersion]{packageVersion()}} if \code{sass_file.scss}
is bundled with an R package).}

\item{options}{Compiler options for Sass. Please specify options using
\code{\link{sass_options}}.}

\item{write_attachments}{If the input contains \code{\link{sass_layer}}
objects that have file attachments, and \code{output} is not \code{NULL},
then copy the file attachments to the directory of \code{output}. (Defaults
to \code{NA}, which merely emits a warning if file attachments are present,
but does not write them to disk; the side-effect of writing extra files is
subtle and potentially destructive, as files may be overwritten.)}

\item{cache}{A \link{FileCache} object created by \code{\link[=sass_file_cache]{sass_file_cache()}}, or \code{NULL}
for no caching.}

\item{script}{Script(s) to include within the document head (should be
specified relative to the \code{src} parameter).}

\item{meta}{Named list of meta tags to insert into document head}

\item{head}{Arbitrary lines of HTML to insert into the document head}
}
\value{
returns a list of \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} objects. The length
of the list use 1 unless \code{input} contains \code{\link[=sass_layer]{sass_layer()}} object(s) that
have \code{html_deps}.
}
\description{
Calls \code{sass(input, ...)} and wraps the output into a \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\examples{
str(sass_html_dependencies(
  input = list(list(color = "red"), "body {color: $color}"),
  name = "foo", version = "1.0.0",
  stylesheet = "my-style.css"
))
}
\seealso{
\code{\link[=sass]{sass()}}
}
