% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{sass_layers}
\alias{sass_layers}
\alias{sass_layer}
\title{Place SASS before and after other SASS}
\usage{
sass_layers(...)

sass_layer(pre = "", post = "", deps = NULL)
}
\arguments{
\item{...}{A collection of \code{\link[=sass_layer]{sass_layer()}}s and/or objects that \code{\link[=as_sass]{as_sass()}} understands.}

\item{pre}{An object that \code{\link[=as_sass]{as_sass()}} understands.}

\item{post}{An object that \code{\link[=as_sass]{as_sass()}} understands.}

\item{deps}{An HTML dependency (or a list of them).}
}
\description{
When combining multiple SASS projects together, it's often desirable to
place SASS files before and after a set of "core" SASS files.
}
\examples{
blue <- list(color = "blue !default")
red <- list(color = "red !default")
green <- list(color = "green !default")

# a sass_layer() by itself is not very useful, it just defines some
# SASS to place before (pre) and after (post)
core <- sass_layer(pre = blue, post = "body { color: $color; }")
core
sass(core)

# However, by composing multiple sass_layers(), we have ability to place
# SASS both before (pre) and after (post) some other sass (e.g., core)
# Here we place a red default _before_ the blue default and export the
# color SASS variable as a CSS variable _after_ the core
red_layer <- sass_layer(red, post = ":root{ --color: #{$color}; }")
sass(sass_layers(core, red_layer))
sass(sass_layers(core, red_layer, sass_layer(green)))

}
