% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{is_sass_removable}
\alias{is_sass_removable}
\alias{sass_removable}
\alias{sass_layer_remove_rule}
\title{Remove Sass layer rule}
\usage{
is_sass_removable(x)

sass_removable(x)

sass_layer_remove_rule(layer, rule)
}
\arguments{
\item{x}{Value to add or check for a \code{"sass_removable"} class}

\item{layer}{Output from \code{\link[=sass_layer]{sass_layer()}}}

\item{rule}{Layer rule name to remove}
}
\description{
Remove Sass layer rule
}
\examples{
# set up a base blue color
blue <- list(color = "blue !default")

# make a layer that has a custom red rule
core <- sass_layer(
  defaults = blue,
  rules = list(
    custom = sass_removable(
      # any R sass definitions can be nested under the `custom` key
      "body > custom { color: $color; }"
    ),
    "body { color: $color; }"
  )
)
# contains custom css
sass(core)

core_slim <- sass_layer_remove_rule(core, "custom")
# does NOT contain custom css
sass(core_slim)
}
