% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{sass_set_rules}
\alias{sass_set_rules}
\title{Set the rules for a \code{sass_layer} object}
\usage{
sass_set_rules(layer, rules)
}
\arguments{
\item{layer}{A \code{\link[=sass_layer]{sass_layer()}} object.}
}
\description{
Replaces the rules for a \code{\link[=sass_layer]{sass_layer()}} object with new rules. This is useful
when (for example) you want to compile a set of rules using variables derived
from a theme, but you do not want the resulting CSS for the entire theme --
just the CSS for the specific rules passed in.
}
\examples{
theme <- sass_layer(
  defaults = sass_file(system.file("examples/variables.scss", package = "sass")),
  rules = sass_file(system.file("examples/rules.scss", package = "sass"))
)

# Compile the theme
sass(theme)

# Sometimes we want to use the variables from the theme to compile other sass
my_rules <- ".someclass { background-color: $bg; color: $fg; }"
sass(sass_set_rules(theme, my_rules))

}
