% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{write_file_attachments}
\alias{write_file_attachments}
\title{Write file attachments from a sass theme object}
\usage{
write_file_attachments(file_attachments, output_path)
}
\arguments{
\item{file_attachments}{A character vector of files or directories.}

\item{output_path}{A directory to copy the attachments to.}
}
\description{
Write file attachments from a sass theme object
}
\keyword{internal}
