% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile .scss and .sass files to CSS}
\usage{
compile(input, ...)
}
\arguments{
\item{input}{path to .scss or .sass file, or string containing
code to compile.}

\item{...}{options to send to libSass. For a full list of supported
options, see the options help page.}
}
\value{
One element character vector containing the compiled CSS.
}
\description{
Compile .scss and .sass files to CSS using libSass.
}
