% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile_sass}
\alias{compile_sass}
\title{Compile Sass to CSS}
\usage{
compile_sass(file = NULL, options = opts(), output = NULL, ...,
  text = NULL)
}
\arguments{
\item{file}{Path to .scss or .sass Sass file.}

\item{options}{Compiler options for Sass. Please specify options using
\code{\link{opts}}.}

\item{output}{Specifies path to output file for compiled CSS.}

\item{...}{Currently unused.}

\item{text}{Character vector of Sass code.}
}
\value{
If \code{output = NULL}, returns a character vector of the compiled
  CSS. If the output path is specified, compiled returns \code{NULL}.
}
\description{
Compile Sass to CSS using libSass.
}
\note{
\code{file} and \code{text} cannot be compiled at the same time. Please
  pass one or the other.
}
\examples{
compile_sass(text = "foo { margin: 122px * .3; }")

}
