% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass.R
\name{output_file}
\alias{output_file}
\title{Create cache aware output file(s)}
\usage{
output_file(basename = "output", pattern = "sass", fileext = NULL)
}
\arguments{
\item{basename}{a non-empty character vector giving the outfile name (without
the extension).}

\item{pattern}{a non-empty character vector giving the initial part of the
directory's basename.}

\item{fileext}{the output file extension. The default is \code{".min.css"} for
compressed and compact output styles; otherwise, its \code{".css"}.}
}
\description{
Create cache aware output file(s)
}
\examples{
sass("body {color: red}", output = output_file())
}
