% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass.R
\name{sass}
\alias{sass}
\title{Compile Sass to CSS}
\usage{
sass(input = NULL, options = sass_options(), output = NULL)
}
\arguments{
\item{input}{Accepts raw Sass, file names, Sass file paths, and/or a named list of variables.
See \code{\link{as_sass}} for more detail.}

\item{options}{Compiler options for Sass. Please specify options using
\code{\link{sass_options}}.}

\item{output}{Specifies path to output file for compiled CSS.}
}
\value{
If \code{output = NULL}, the function returns a string value
  of the compiled CSS. If the output path is specified, the compiled
  CSS is written to that file.
}
\description{
Compile Sass to CSS using LibSass.
}
\examples{
sass("foo { margin: 122px * .3; }")
sass(list(
  list(width = "122px"),
  "foo { margin: $width * .3; }"
))
}
\seealso{
\url{http://sass-lang.com/guide}
}
