% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{sass}
\alias{sass}
\title{Compile Sass to CSS}
\usage{
sass(input = NULL, options = sass_options(), output = NULL)
}
\arguments{
\item{input}{Raw Sass text or a file path to .scss or .sass Sass file.
Note that the libSass compiler expects .sass files to use the Sass Indented Syntax.}

\item{options}{Compiler options for Sass. Please specify options using
\code{\link{sass_options}}.}

\item{output}{Specifies path to output file for compiled CSS.}
}
\value{
If \code{output = NULL}, the function returns a one element character
  vector of the compiled CSS. If the output path is specified, the compiled
  CSS is written to that file and the function has no return value.
}
\description{
Compile Sass to CSS using libSass.
}
\examples{
sass("foo { margin: 122px * .3; }")

}
