% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass_cache.R
\name{sass_cache_get}
\alias{sass_cache_get}
\alias{sass_cache_set}
\title{Get and set the cache object used for a specific directory}
\usage{
sass_cache_get(dir)

sass_cache_set(dir, cache)
}
\arguments{
\item{dir}{A directory. If \code{NULL}, this function will return \code{NULL}.}

\item{cache}{A \code{\link[=sass_file_cache]{sass_file_cache()}} object, or \code{NULL} if you don't want to use
a cache.}
}
\description{
Get and set the cache object used for a specific directory
}
\details{
If \code{sass_cache_get()} is called for a given directory, before
\code{sass_cache_set()} has been called for that directory, then a new
\code{\link[=sass_file_cache]{sass_file_cache()}} object will be created and set for that directory.

After \code{sass_cache_set()} is called for a directory, any future calls to
\code{sass_cache_get()} with that directory will return that specific cache
object. This can be useful if you customize parameters for the cache object,
like maximum size or age.
}
\seealso{
\code{\link[=sass_file_cache]{sass_file_cache()}}, \code{\link[=sass]{sass()}}
}
\keyword{internal}
