% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass_cache.R
\name{sass_cache_get}
\alias{sass_cache_get}
\title{Retrieve the default file cache}
\usage{
sass_cache_get()
}
\description{
When caching is enabled, this function returns a \code{\link[=sass_file_cache]{sass_file_cache()}} object
that \code{\link[=sass]{sass()}}'s \code{cache} argument uses (by default) for caching Sass
compilation. When caching is disabled (either by setting the \code{sass.cache}
option to \code{FALSE}, \code{NULL}, or via \code{\link[shiny:devmode]{shiny::devmode()}}), this function returns
\code{NULL} (effectively telling \code{\link[=sass]{sass()}} to not \code{cache} by default).
}
\details{
When caching is enabled, then this function returns a \code{sass_file_cache()}
object that (by default) uses \code{\link[=sass_cache_context_dir]{sass_cache_context_dir()}} for its directory.
The directory can also be customized by providing the \code{sass.cache} option
with either a filepath (as a string) or a full-blown \code{sass_file_cache()}
object.
}
\seealso{
\code{\link[=sass_cache_get_dir]{sass_cache_get_dir()}}, \code{\link[=sass]{sass()}}
}
