% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sass_cache.R
\name{sass_cache_options}
\alias{sass_cache_options}
\title{Caching Options for Sass (defunct)}
\usage{
sass_cache_options(cache, cache_dir)
}
\arguments{
\item{cache}{No longer used.}

\item{cache_dir}{No longer used.}
}
\description{
This function is no longer used. Please see \code{\link[=sass_cache_get]{sass_cache_get()}}.
}
\keyword{internal}
